added Client class
added the Client class which creates a client for the user to user when
the IP, port and optionally name are inputted. The name can be put in
but if it isn't the Namer GUI will be run. The client when started will
perform a handshake with the server where they pass their Public keys to
each other and verify them, this will allow to to start encrypting the
messages sent between the users by using AES to encrypt the message and
RSA to encrypt the AES key.